+++ src/lib/include/AST/values/ADLDecimal.hpp
@@ -4,7 +4,7 @@
  * Created:
  *   25/11/2020, 16:13:03
  * Last edited:
- *   27/11/2020, 14:54:55
+ *   12/2/2020, 2:47:12 PM
  * Auto updated?
  *   Yes
  *
@@ -32,7 +32,7 @@ namespace ArgumentParser {
         double value;
 
         /* Constructor for the ADLDecimal class, which takes a breadcrumb trail of filenames where this node originated from, a DebugInfo struct linking this node to a location in the source file and the value this node represents. */
-        ADLDecimal(const std::vector<std::string>& filenames, const DebugInfo& debug, long value);
+        ADLDecimal(const std::vector<std::string>& filenames, const DebugInfo& debug, double value);
 
         /* Prints the decimal to the given output stream. */
         virtual std::ostream& print(std::ostream& os) const;
